+++ terra_geocrud/tests/test_validators.py
@@ -1,7 +1,20 @@
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from terra_geocrud.validators import validate_schema_property
+from terra_geocrud.validators import validate_schema_property, validate_function_path
+
+
+class ValidateFunctionPathTestCase(TestCase):
+    def test_validator_with_wrong_package(self):
+        with self.assertRaises(ValidationError):
+            validate_function_path('not_available')
+
+    def test_validator_with_wrong_function_path(self):
+        with self.assertRaises(ValidationError):
+            validate_function_path('test_terra_geocrud.functions_test.not_available')
+
+    def test_validator_with_right_function_path(self):
+        validate_function_path('test_terra_geocrud.functions_test.get_length')
 
 
 class ValidateSchemaPropertyTestCase(TestCase):