+++ terra_geocrud/validators.py
@@ -1,3 +1,6 @@
+from django.core.exceptions import ValidationError
+from django.utils.module_loading import import_string
+
 from geostore.validators import validate_json_schema
 
 
@@ -9,3 +12,12 @@ def validate_schema_property(value):
         }
     }
     return validate_json_schema(json_schema)
+
+
+def validate_function_path(value):
+    if value:
+        try:
+            import_string(value)
+        except ImportError:
+            raise ValidationError(message=f"function {value} does not exist")
+    return value